fix: use getPlatformVersion to filter platformApps Moved the filter from the old provider folder and cleanup old code
