+++ libs/ledger-live-common/src/platform/filters.ts
@@ -1,6 +1,14 @@
 import { makeRe } from "minimatch";
-import { PlatformAccount, PlatformCurrency } from "./types";
+import { PlatformAccount, PlatformCurrency, AppManifest } from "./types";
 import { isPlatformTokenCurrency } from "./helpers";
+import semver from "semver";
+
+export type FilterParams = {
+  branches?: string[];
+  platform?: string;
+  private?: boolean;
+  version?: string;
+};
 
 export type AccountFilters = {
   currencies?: string[];
@@ -51,5 +59,47 @@ export function filterPlatformCurrencies(
     }
 
     return true;
+  });
+}
+
+function matchVersion(filterParams: FilterParams, manifest: AppManifest) {
+  return (
+    !filterParams.version ||
+    semver.satisfies(
+      semver.coerce(filterParams.version) || "",
+      manifest.apiVersion
+    )
+  );
+}
+
+function matchBranches(filterParams: FilterParams, manifest: AppManifest) {
+  return (
+    !filterParams.branches || filterParams.branches.includes(manifest.branch)
+  );
+}
+
+function matchPlatform(filterParams: FilterParams, manifest: AppManifest) {
+  return (
+    !filterParams.platform ||
+    manifest.platform === "all" ||
+    filterParams.platform === manifest.platform
+  );
+}
+
+function matchPrivate(filterParams: FilterParams, manifest: AppManifest) {
+  return filterParams.private === true || !(manifest.private === true);
+}
+
+export function filterPlatformApps(
+  appManifests: AppManifest[],
+  filterParams: FilterParams
+): AppManifest[] {
+  return appManifests.filter((appManifest: AppManifest) => {
+    return (
+      matchBranches(filterParams, appManifest) &&
+      matchPlatform(filterParams, appManifest) &&
+      matchPrivate(filterParams, appManifest) &&
+      matchVersion(filterParams, appManifest)
+    );
   });
 }
